feat: added auth to project routes

+++ backend/src/api/routes/projects.ts
@@ -4,6 +4,8 @@ import { AppError } from '../errors/AppError.js'
 import type { DbError } from '../errors/DbError.js'
 import dbErrorMapper from '../errors/dbErrorMapper.js'
 import {
+  validateProjectDelete,
+  validateProjectGet,
   validateProjectPatch,
   validateProjectPost,
 } from '../validators/projects_validation.js'
@@ -11,11 +13,20 @@ import {
   buildProjectPatchQuery,
   type projectPatchReqBody,
 } from '../queries/projectPatchQuery.js'
+import type { AuthenticatedRequest } from '../../types/authenticatedRequest.js'
+import {
+  getProject,
+  isProjectMember,
+  isProjectOwner,
+} from '../../db/services/project.services.js'
 
 const projectRouter = Router()
 
 // Create new project
-projectRouter.post('/', async (req, res) => {
+projectRouter.post('/', async (req: AuthenticatedRequest, res) => {
+  const user = req.user
+  if (!user) throw new AppError('UNAUTHORIZED_REQUEST')
+
   validateProjectPost(req.body)
 
   const text =
@@ -23,7 +34,7 @@ projectRouter.post('/', async (req, res) => {
   const values = [
     req.body.name,
     req.body.description ?? null,
-    req.body.owner_id,
+    user.sub,
     req.body.code,
   ]
 
@@ -35,17 +46,26 @@ projectRouter.post('/', async (req, res) => {
   }
 })
 
-// Get existing project
-projectRouter.get('/:id', async (req, res) => {
-  const text = 'SELECT * FROM projects WHERE id = $1'
-  const values = [req.params.id]
+projectRouter.get('/:id', async (req: AuthenticatedRequest, res) => {
+  const { user, projectId } = validateProjectGet(req.params.id, req.user)
+  const authorized = await isProjectMember(projectId, user.sub)
+
+  const text = `SELECT * FROM projects WHERE id = $1` // be owner or contributor
+  const values = [projectId]
 
   try {
     const result = await pool.query(text, values)
-    if (result.rowCount === 0) {
+
+    // Differentiating despite security cost because this is a portfolio/practice project
+    if (result.rowCount !== 0) {
+      if (authorized) {
+        return res.status(200).json(result.rows[0])
+      } else {
+        throw new AppError('UNAUTHORIZED_REQUEST')
+      }
+    } else {
       throw new AppError('PROJECT_NOT_FOUND')
     }
-    return res.status(200).json(result.rows[0])
   } catch (err) {
     dbErrorMapper(err as DbError)
   }
@@ -68,12 +88,12 @@ projectRouter.get('/', async (req, res) => {
   }
 })
 
-projectRouter.patch('/:id', async (req, res) => {
-  validateProjectPatch(req.body)
+projectRouter.patch('/:id', async (req: AuthenticatedRequest, res) => {
+  const { id, user } = validateProjectPatch(req.params.id, req.body, req.user)
 
   const [text, values] = buildProjectPatchQuery(
     req.body as projectPatchReqBody,
-    req.params.id,
+    id,
   )
 
   try {
@@ -81,25 +101,33 @@ projectRouter.patch('/:id', async (req, res) => {
     if (result.rowCount === 0) {
       throw new AppError('PROJECT_NOT_FOUND')
     }
+    await isProjectOwner(id, user.sub)
     return res.status(200).json(result.rows[0])
   } catch (err) {
     dbErrorMapper(err as DbError)
   }
 })
 
-projectRouter.delete('/:id', async (req, res) => {
+projectRouter.delete('/:id', async (req: AuthenticatedRequest, res) => {
+  const { id, user } = validateProjectDelete(req.params.id, req.user)
+
   const text = `
     DELETE FROM projects
     WHERE id = $1
+    AND owner_id = $2
     RETURNING *
     `
 
-  const values = [req.params.id]
+  const values = [id, user.sub]
 
   try {
     const result = await pool.query(text, values)
     if (result.rowCount === 0) {
-      throw new AppError('PROJECT_NOT_FOUND')
+      if (await getProject(id)) {
+        throw new AppError('UNAUTHORIZED_REQUEST')
+      } else {
+        throw new AppError('PROJECT_NOT_FOUND')
+      }
     }
     return res.status(204).send()
   } catch (err) {
